Code refactoring for frontend
