+++ Travlendar/frontend/src/Services/SignupService.js
@@ -1,4 +1,12 @@
 import axios from 'axios';
+
+/*
+    Author: Kanchan Wakchaure
+    Date: 10-05-2019
+    Description: Signup service to post user details
+    References: https://material-ui.com
+*/
+
 const API_URL = 'http://localhost:8000';
 
 export default class SignupService{
