+++ Travlendar/frontend/src/components/home/Homepage.js
@@ -1,11 +1,5 @@
-import React, { Component } from "react";
-import { BrowserRouter as Router, Route,Switch, Redirect} from 'react-router-dom';
+import React from "react";
 import './Homepage.css'
-import Login from './login';
-import Signup from './register/Signup';
-import CreateEvent from './CreateEvent';
-import Container from '@material-ui/core/Container';
-import Calendar from 'react-calendar';
 
 class Homepage extends React.Component{
     render() {
