+++ Travlendar/users/urls.py
@@ -3,5 +3,5 @@ from django.urls import include, path
 from . import views
 
 urlpatterns = [
-    path('signup/', views.signUp),
+    path('register/', views.signUp),
 ]
